requirements.txt
openai>=1.40.0
pandas>=2.2.0
pyyaml>=6.0.1
python-dotenv>=1.0.1
tenacity>=8.4.2
tqdm>=4.66.4
pytest>=8.2.0
